geo_toolz/_src/crs/test_coords.py
import numpy as np
from geoprocess.crs.coords import convert_x_y_to_lat_lon, convert_lat_lon_to_x_y, calc_latlon


def test_calc_latlon():
    # Create a sample dataset with x and y coordinates
    x = np.arange(0, 10)
    y = np.arange(0, 5)
    data = np.random.rand(len(y), len(x))
    ds = xr.Dataset({"data": (["y", "x"], data)}, coords={"x": x, "y": y})

    # Call the function to calculate latitude and longitude coordinates
    result = calc_latlon(ds)

    # Check if the latitude and longitude coordinates are added to the dataset
    assert "latitude" in result.coords
    assert "longitude" in result.coords

    # Check if the latitude and longitude arrays have the correct shape
    assert result.latitude.shape == (len(y), len(x))
    assert result.longitude.shape == (len(y), len(x))

    # Check if the latitude and longitude arrays have the correct units
    assert result.latitude.attrs["units"] == "degrees_north"
    assert result.longitude.attrs["units"] == "degrees_east"import numpy as np


def test_calc_latlon():
    # Create a sample dataset with x and y coordinates
    x = np.arange(0, 10)
    y = np.arange(0, 5)
    data = np.random.rand(len(y), len(x))
    ds = xr.Dataset({"data": (["y", "x"], data)}, coords={"x": x, "y": y})

    # Call the function to calculate latitude and longitude coordinates
    result = calc_latlon(ds)

    # Check if the latitude and longitude coordinates are added to the dataset
    assert "latitude" in result.coords
    assert "longitude" in result.coords

    # Check if the latitude and longitude arrays have the correct shape
    assert result.latitude.shape == (len(y), len(x))
    assert result.longitude.shape == (len(y), len(x))

    # Check if the latitude and longitude arrays have the correct units
    assert result.latitude.attrs["units"] == "degrees_north"
    assert result.longitude.attrs["units"] == "degrees_east"


def test_convert_x_y_to_lat_lon():
    # Define test coordinates and CRS
    x = [0, 1, 2, 3, 4]
    y = [0, 1, 2, 3, 4]
    crs = "epsg:3857"

    # Call the function to convert x and y coordinates to latitude and longitude
    lon, lat = convert_x_y_to_lat_lon(crs, x, y)

    # Check if the returned longitude and latitude values are of type float
    assert isinstance(lon, float)
    assert isinstance(lat, float)

    # Add more assertions to validate the correctness of the conversion
    # ...

    # Add more test cases to cover different scenarios
    # ...import numpy as np


def test_convert_lat_lon_to_x_y():
    # Define test coordinates and CRS
    lon = [0, 1, 2, 3, 4]
    lat = [0, 1, 2, 3, 4]
    crs = "epsg:3857"

    # Call the function to convert latitude and longitude coordinates to x and y
    x, y = convert_lat_lon_to_x_y(crs, lon, lat)

    # Check if the returned x and y values are of type float
    assert isinstance(x, float)
    assert isinstance(y, float)

    # Add more assertions to validate the correctness of the conversion
    # ...

    # Add more test cases to cover different scenarios
    # ...

def test_calc_latlon():
    # Create a sample dataset with x and y coordinates
    x = np.arange(0, 10)
    y = np.arange(0, 5)
    data = np.random.rand(len(y), len(x))
    ds = xr.Dataset({"data": (["y", "x"], data)}, coords={"x": x, "y": y})

    # Call the function to calculate latitude and longitude coordinates
    result = calc_latlon(ds)

    # Check if the latitude and longitude coordinates are added to the dataset
    assert "latitude" in result.coords
    assert "longitude" in result.coords

    # Check if the latitude and longitude arrays have the correct shape
    assert result.latitude.shape == (len(y), len(x))
    assert result.longitude.shape == (len(y), len(x))

    # Check if the latitude and longitude arrays have the correct units
    assert result.latitude.attrs["units"] == "degrees_north"
    assert result.longitude.attrs["units"] == "degrees_east"

def test_convert_x_y_to_lat_lon():
    # Define test coordinates and CRS
    x = [0, 1, 2, 3, 4]
    y = [0, 1, 2, 3, 4]
    crs = "epsg:3857"

    # Call the function to convert x and y coordinates to latitude and longitude
    lon, lat = convert_x_y_to_lat_lon(crs, x, y)

    # Check if the returned longitude and latitude values are of type float
    assert isinstance(lon, float)
    assert isinstance(lat, float)

    # Add more assertions to validate the correctness of the conversion
    # ...

    # Add more test cases to cover different scenarios
    # ...